[#8] Feat: 추가 정보 옵션 파일로 분리
