+++ src/pages/SignInPage/OptionalInfo.tsx
@@ -1,43 +1,11 @@
 import { useNavigate } from 'react-router-dom';
 import InfoSelect from '../../components/InfoSelect';
+import age from '../../utils/age';
+import banks from '../../utils/banks';
+import jobs from '../../utils/jobs';
+import regions from '../../utils/regions';
 
 type Props = {};
-const jobs = ['회사원', '재직자', '자영업자', '대학생', '군인', '무직'];
-const regions = [
-  '전국',
-  '서울',
-  '경기',
-  '인천',
-  '부산',
-  '광주',
-  '대구',
-  '대전',
-  '울산',
-  '경북',
-  '전북',
-  '제주',
-];
-
-const age: number[] = [];
-for (let i = 20; i < 70; i++) {
-  age.push(i);
-}
-
-const banks = [
-  'KEB하나은행',
-  'SC제일은행',
-  '국민은행',
-  '신한은행',
-  '외환은행',
-  '우리은행',
-  '한국시티은행',
-  '기업은행',
-  '농협',
-  '수협',
-  '한국산업은행',
-  '부산은행',
-  '전북은행',
-];
 
 const OptionalInfo = (props: Props) => {
   const navigate = useNavigate();
