+++ src/utils/age.ts
@@ -1,0 +1,6 @@
+const age: number[] = [];
+for (let i = 20; i < 70; i++) {
+  age.push(i);
+}
+
+export default age;
