+++ src/utils/jobs.ts
@@ -1,0 +1,3 @@
+const jobs = ['대학생', '회사원', '자영업자', '무직'];
+
+export default jobs;
