+++ src/utils/regions.ts
@@ -1,0 +1,22 @@
+const regions = [
+  '전국',
+  '서울',
+  '경기',
+  '인천',
+  '부산',
+  '광주',
+  '대구',
+  '대전',
+  '울산',
+  '세종',
+  '강원',
+  '전남',
+  '전북',
+  '충남',
+  '충북',
+  '경남',
+  '경북',
+  '제주',
+];
+
+export default regions;
